pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "socdata"
version = "0.1.0"
description = "Multi-dataset sociology research assistant powered by Claude AI"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "python-dotenv>=1.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "statsmodels>=0.14",
    "scipy>=1.10",
    "pyreadstat>=1.2",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
census = ["census>=0.8", "us>=3.0"]
ipums = ["ipumspy>=0.4"]
all = ["socdata[census,ipums]"]

[project.scripts]
socdata = "socdata.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["socdata*"]
